Add files via upload
A program dedicated to simulating the Mathematical Game of 50 where the user can choose Player 1's strategy. The choices are to choose a random number or to choose a strategy that is guaranteed to win every time. When the program starts, input 0 to choose the random strategy and input 1 to choose the winning strategy. Then input the number of game iterations you wish for the program to run. Press enter after each of your inputs. Once the results have been printed, you may scroll up or down to see them. Input y to restart the program or n to end it.